refactor data distributions page to display demo visualizations
