+++ website/src/app/pages/data-distributions/data-distributions.component.html
@@ -6,22 +6,19 @@
   (dataVariableChange)="data.setCurrentDataVariable($event)"
 ></agc-table-data-selector>
 
-<!-- <div class="data-block" *ngIf="!!$any(data.currentSpec$ | async)">
-  <agc-dataset-summary
-    [summary]="data.currentColumnSummary$ | async"
-    [title]="data.currentDataVariable$ | async"
-  ></agc-dataset-summary>
+<div class="data-block">
   <ngx-vega
-    *ngIf="$any(data.currentColumn$ | async).distType !== 'summary'"
-    [spec]="$any(data.currentSpec$ | async)"
+    [spec]="demoPieSpec"
     [options]="{}"
     [autosize]="{ width: true, height: false }"
   ></ngx-vega>
-</div> -->
-
-<div class="data-block">
   <ngx-vega
-    [spec]="spec"
+    [spec]="demoBarChartSpec"
+    [options]="{}"
+    [autosize]="{ width: true, height: false }"
+  ></ngx-vega>
+  <ngx-vega
+    [spec]="demoHorizBarChartSpec"
     [options]="{}"
     [autosize]="{ width: true, height: false }"
   ></ngx-vega>
